trending-data-collection/requirements.txt
google-auth==2.23.4
google-api-python-client==2.108.0
python-dotenv==1.0.0
browserbase
playwright
requests
tavily-python
schedule
